CLN: remove legacy periods (#182)
(cherry picked from commit 8ecb86bcff8204a609a5ef86e71feda9dd905104)
