+++ python/rateslib/instruments/components/fx_options/call_put.py
@@ -24,7 +24,7 @@ from rateslib.instruments.components.protocols.pricing import (
 )
 from rateslib.legs.components import CustomLeg
 from rateslib.periods.components import Cashflow, FXCallPeriod, FXPutPeriod
-from rateslib.periods.utils import _validate_fx_as_forwards
+from rateslib.periods.components.utils import _validate_fx_as_forwards
 from rateslib.scheduling.frequency import _get_fx_expiry_and_delivery_and_payment
 
 if TYPE_CHECKING:
