+++ python/rateslib/periods/__init__.py
@@ -1,25 +1,21 @@
-from rateslib.periods.base import BasePeriod
-from rateslib.periods.cashflow import Cashflow, NonDeliverableCashflow
-from rateslib.periods.credit import CreditPremiumPeriod, CreditProtectionPeriod
-from rateslib.periods.fx_volatility import FXCallPeriod, FXOptionPeriod, FXPutPeriod
-from rateslib.periods.index import IndexCashflow, IndexFixedPeriod, IndexMixin
-from rateslib.periods.rates import FixedPeriod, FloatPeriod, NonDeliverableFixedPeriod
-
-# from rateslib.periods.components import FixedPeriod
+from rateslib.periods.components import (
+    Cashflow,
+    CreditPremiumPeriod,
+    CreditProtectionPeriod,
+    FixedPeriod,
+    FloatPeriod,
+    FXCallPeriod,
+    FXOptionPeriod,
+    FXPutPeriod,
+)
 
 __all__ = [
-    "BasePeriod",
     "FXOptionPeriod",
     "FXPutPeriod",
     "FXCallPeriod",
     "Cashflow",
-    "NonDeliverableCashflow",
     "CreditPremiumPeriod",
     "CreditProtectionPeriod",
-    "IndexCashflow",
-    "IndexFixedPeriod",
     "FixedPeriod",
-    "NonDeliverableFixedPeriod",
     "FloatPeriod",
-    "IndexMixin",
 ]
